+++ assets/js/admin/woocommerce_admin.js
@@ -406,6 +406,43 @@
 				return window.confirm( woocommerce_admin.i18n_remove_personal_data_notice );
 			}
 		});
+
+		var marketplaceSectionDropdown = $( '#marketplace-current-section-dropdown' );
+		var marketplaceSectionName = $( '#marketplace-current-section-name' );
+		var marketplaceMenuIsOpen = false;
+
+		// Add event listener to toggle Marketplace menu on touch devices
+		if ( marketplaceSectionDropdown.length && isTouchDevice() ) {
+			marketplaceSectionName.on( 'click', function() {
+				marketplaceMenuIsOpen = ! marketplaceMenuIsOpen;
+				if ( marketplaceMenuIsOpen ) {
+					marketplaceSectionDropdown.addClass( 'is-open' );
+					$( document ).on( 'click', maybeToggleMarketplaceMenu );
+				} else {
+					marketplaceSectionDropdown.removeClass( 'is-open' );
+					$( document ).off( 'click', maybeToggleMarketplaceMenu );
+				}
+			} );
+		}
+
+		// Close menu if the user clicks outside it
+		function maybeToggleMarketplaceMenu( e ) {
+			if (
+				! marketplaceSectionDropdown.is( e.target )
+				&& marketplaceSectionDropdown.has( e.target ).length === 0
+			) {
+				marketplaceSectionDropdown.removeClass( 'is-open' );
+				marketplaceMenuIsOpen = false;
+				$( document ).off( 'click', maybeToggleMarketplaceMenu );
+			}
+		}
+
+		function isTouchDevice() {
+			return ( ( 'ontouchstart' in window ) ||
+				( navigator.maxTouchPoints > 0 ) ||
+				( navigator.msMaxTouchPoints > 0 ) );
+		}
+
 	});
 
 })( jQuery, woocommerce_admin );