+++ includes/class-wc-discounts.php
@@ -84,7 +84,7 @@ class WC_Discounts {
 			$item->object        = $cart_item;
 			$item->product       = $cart_item['data'];
 			$item->quantity      = $cart_item['quantity'];
-			$item->price         = wc_add_number_precision_deep( $item->product->get_price() * $item->quantity );
+			$item->price         = wc_add_number_precision_deep( (float) $item->product->get_price() * (float) $item->quantity );
 			$this->items[ $key ] = $item;
 		}