+++ tests/e2e/config/jest.setup.js
@@ -5,12 +5,13 @@ import {
 	WP_ADMIN_LOGIN
 } from '@woocommerce/e2e-utils';
 
-const config = require('config');
-const { HTTPClientFactory } = require('@woocommerce/api');
+const config = require( 'config' );
+const { HTTPClientFactory } = require( '@woocommerce/api' );
 const { addConsoleSuppression, updateReadyPageStatus } = require( '@woocommerce/e2e-environment' );
+const { DEFAULT_TIMEOUT_OVERRIDE } = process.env;
 
 // @todo: remove this once https://github.com/woocommerce/woocommerce-admin/issues/6992 has been addressed
-addConsoleSuppression( 'woocommerce_shared_settings' );
+addConsoleSuppression( 'woocommerce_shared_settings', false );
 
 /**
  * Uses the WordPress API to delete all existing posts
@@ -38,12 +39,23 @@ async function trashExistingPosts() {
 // other posts/comments/etc. aren't dirtying tests and tests don't depend on
 // each other's side-effects.
 beforeAll(async () => {
-	// Update the ready page to prevent concurrent test runs
-	await updateReadyPageStatus('draft');
 
-	await trashExistingPosts();
-	await withRestApi.deleteAllProducts();
-	await withRestApi.deleteAllCoupons();
+	if ( DEFAULT_TIMEOUT_OVERRIDE ) {
+		page.setDefaultNavigationTimeout( DEFAULT_TIMEOUT_OVERRIDE );
+		page.setDefaultTimeout( DEFAULT_TIMEOUT_OVERRIDE );
+	}
+
+	try {
+		// Update the ready page to prevent concurrent test runs
+		await updateReadyPageStatus('draft');
+		await trashExistingPosts();
+		await withRestApi.deleteAllProducts();
+		await withRestApi.deleteAllCoupons();
+		await withRestApi.deleteAllOrders();
+	} catch ( error ) {
+		// Prevent an error here causing tests to fail.
+	}
+
 	await page.goto(WP_ADMIN_LOGIN);
 	await clearLocalStorage();
 	await setBrowserViewport('large');
@@ -53,7 +65,11 @@ beforeAll(async () => {
 // This is to ensure that each test ends with no user logged in.
 afterAll(async () => {
 	// Reset the ready page to published to allow future test runs
-	await updateReadyPageStatus('publish');
+	try {
+		await updateReadyPageStatus('publish');
+	} catch ( error ) {
+		// Prevent an error here causing tests to fail.
+	}
 
 	const client = await page.target().createCDPSession();
 	await client.send('Network.clearBrowserCookies');