+++ tests/e2e/core-tests/specs/merchant/wp-admin-extensions-connect-wccom.test.js
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-export, jest/no-disabled-tests */
 /**
  * Internal dependencies
  */
@@ -22,13 +21,7 @@ const runInitiateWccomConnectionTest = () => {
 		});
 
 		it('can initiate WCCOM connection', async () => {
-			await merchant.openExtensions();
-
-			// Click on a tab to choose WooCommerce Subscriptions extension
-			await Promise.all([
-				expect(page).toClick('a.nav-tab', {text: "WooCommerce.com Subscriptions"}),
-				page.waitForNavigation({waitUntil: 'networkidle0'}),
-			]);
+			await merchant.openHelper();
 
 			// Click on Connect button to initiate a WCCOM connection
 			await Promise.all([