+++ tests/e2e/core-tests/specs/merchant/wp-admin-product-new.test.js
@@ -1,4 +1,3 @@
-/* eslint-disable jest/no-export, jest/no-disabled-tests */
 /**
  * Internal dependencies
  */
@@ -8,7 +7,9 @@ const {
 	clickTab,
 	uiUnblocked,
 	evalAndClick,
+	setCheckbox,
 	verifyAndPublish,
+	waitForSelectorWithoutThrow
 } = require( '@woocommerce/e2e-utils' );
 const {
 	waitAndClick,
@@ -28,26 +29,6 @@ const config = require( 'config' );
 const VirtualProductName = 'Virtual Product Name';
 const NonVirtualProductName = 'Non-Virtual Product Name';
 const simpleProductPrice = config.has('products.simple.price') ? config.get('products.simple.price') : '9.99';
-
-const verifyPublishAndTrash = async () => {
-	// Wait for auto save
-	await page.waitFor( 2000 );
-
-	// Publish product
-	await expect( page ).toClick( '#publish' );
-	await page.waitForSelector( '.updated.notice', { text: 'Product published.' } );
-
-	// Verify
-	await expect( page ).toMatchElement( '.updated.notice', { text: 'Product published.' } );
-	await page.waitForSelector( 'a', { text: 'Move to Trash' } );
-
-	// Trash product
-	await expect( page ).toClick( 'a', { text: 'Move to Trash' } );
-	await page.waitForSelector( '.updated.notice', { text: '1 product moved to the Trash.' } );
-
-	// Verify
-	await expect( page ).toMatchElement( '.updated.notice', { text: '1 product moved to the Trash.' } );
-};
 
 const openNewProductAndVerify = async () => {
 	// Go to "add product" page
@@ -74,7 +55,9 @@ const runAddSimpleProductTest = () => {
 			await verifyAndPublish();
 
 			await merchant.logout();
+		});
 
+		it('can have a shopper add the simple virtual product to the cart', async () => {
 			// See product in the shop and add it to the cart
 			await shopper.goToShop();
 			await shopper.addToCartFromShopPage(VirtualProductName);
@@ -99,9 +82,14 @@ const runAddSimpleProductTest = () => {
 			await verifyAndPublish();
 
 			await merchant.logout();
+		});
 
+		it('can have a shopper add the simple non-virtual product to the cart', async () => {
 			// See product in the shop and add it to the cart
 			await shopper.goToShop();
+
+			await page.reload({ waitUntil: ['networkidle0', 'domcontentloaded'] });
+
 			await shopper.addToCartFromShopPage(NonVirtualProductName);
 			await shopper.goToCart();
 			await shopper.productIsInCart(NonVirtualProductName);
@@ -174,7 +162,7 @@ const runAddVariableProductTest = () => {
 			await uiUnblocked();
 		});
 
-		it('can create variation attributes', async () => {
+		it('can add variation attributes', async () => {
 			await waitAndClick( page, '.variations_tab a' );
 			await uiUnblocked();
 			await waitForSelector(
@@ -187,46 +175,24 @@ const runAddVariableProductTest = () => {
 			);
 
 			// Verify that variations were created
-			await Promise.all([
-				expect(page).toMatchElement('select[name="attribute_attr-1[0]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[0]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[0]"]', {text: 'val1'}),
+			for ( let index = 0; index < 8; index++ ) {
+				const val1 = { text: 'val1' };
+				const val2 = { text: 'val2' };
 
-				expect(page).toMatchElement('select[name="attribute_attr-1[1]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[1]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[1]"]', {text: 'val2'}),
+				// odd / even
+				let attr3 = !! ( index % 2 );
+				// 0-1,4-5 / 2-3,6-7
+				let attr2 = ( index % 4 ) > 1;
+				// 0-3 / 4-7
+				let attr1 = ( index > 3 );
 
-				expect(page).toMatchElement('select[name="attribute_attr-1[2]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[2]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[2]"]', {text: 'val1'}),
+				await expect( page ).toMatchElement( `select[name="attribute_attr-1[${index}]"]`, attr1 ? val2 : val1 );
+				await expect( page ).toMatchElement( `select[name="attribute_attr-2[${index}]"]`, attr2 ? val2 : val1 );
+				await expect( page ).toMatchElement( `select[name="attribute_attr-3[${index}]"]`, attr3 ? val2 : val1 );
+			}
 
-				expect(page).toMatchElement('select[name="attribute_attr-1[3]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[3]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[3]"]', {text: 'val2'}),
-
-				expect(page).toMatchElement('select[name="attribute_attr-1[4]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[4]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[4]"]', {text: 'val1'}),
-
-				expect(page).toMatchElement('select[name="attribute_attr-1[5]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[5]"]', {text: 'val1'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[5]"]', {text: 'val2'}),
-
-				expect(page).toMatchElement('select[name="attribute_attr-1[6]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[6]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[6]"]', {text: 'val1'}),
-
-				expect(page).toMatchElement('select[name="attribute_attr-1[7]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-2[7]"]', {text: 'val2'}),
-				expect(page).toMatchElement('select[name="attribute_attr-3[7]"]', {text: 'val2'}),
-			]);
-
-			/*
-			Puppeteer seems unable to find the individual variation fields in headless mode on MacOS
-			This section of the test runs fine in both Travis and non-headless mode on Mac
-			Disabling temporarily to allow the test to be re-enabled without local testing headache
 			await waitAndClick( page, '.variations-pagenav .expand_all');
-			await page.waitFor( 2000 );
+			await waitForSelectorWithoutThrow( 'input[name="variable_is_virtual[0]"]', 5 );
 			await setCheckbox('input[name="variable_is_virtual[0]"]');
 			await expect(page).toFill('input[name="variable_regular_price[0]"]', '9.99');
 
@@ -242,15 +208,7 @@ const runAddVariableProductTest = () => {
 
 			await page.focus('button.save-variation-changes');
 			await expect(page).toClick('button.save-variation-changes', {text: 'Save changes'});
-			/**/
-
-			// Publish product, verify that it was published. Trash product, verify that it was trashed.
-			await verifyPublishAndTrash(
-				'#publish',
-				'.notice',
-				'Product published.',
-				'1 product moved to the Trash.'
-			);
+			// @todo: https://github.com/woocommerce/woocommerce/issues/30580
 		});
 	});
 };