+++ tests/e2e/env/utils/get-plugin-zip.js
@@ -30,6 +30,52 @@ const getRemotePluginZip = async ( fileUrl ) => {
 };
 
 /**
+ * Get the latest release zip for a plugin from a GiHub repository.
+ *
+ * @param {string} owner The owner of the plugin repository.
+ * @param {string} repository The repository name.
+ * @param {boolean} getPrerelease Flag on whether to get a prelease or not.
+ * @param {number} perPage Limit of entries returned from the latest releases list, defaults to 3.
+ * @returns {Promise<string>}} Returns the URL for the release zip file.
+ */
+const getLatestReleaseZipUrl = async ( owner, repository, getPrerelease = false, perPage = 3 ) => {
+	let requesturl;
+
+	if ( getPrerelease ) {
+		requesturl = `https://api.github.com/repos/${owner}/${repository}/releases?per_page=${perPage}`;
+	} else {
+		requesturl = `https://api.github.com/repos/${owner}/${repository}/releases/latest`;
+	}
+
+	const options = {
+		url: requesturl,
+		method: 'GET',
+		json: true,
+		headers: {'user-agent': 'node.js'}
+	};
+
+	// Wrap in a promise to make the request async
+	return new Promise( function( resolve, reject ) {
+		request.get(options, function( err, resp, body ) {
+			if ( err ) {
+				reject( err );
+			} else {
+				if ( getPrerelease ) {
+					// Loop until we find the first pre-release, then return it.
+					body.forEach(release => {
+						if ( release.prerelease ) {
+							resolve( release.assets[0].browser_download_url );
+						}
+					});
+				} else {
+					resolve( body.assets[0].browser_download_url );
+				}
+			}
+		})
+	});
+}
+
+/**
  * Check the zip file for any nested zips. If one is found, extract it.
  *
  * @param {string} zipFilePath The location of the zip file.
@@ -80,6 +126,7 @@ const downloadZip = async ( fileUrl, downloadPath ) => {
 
 module.exports = {
 	getRemotePluginZip,
+	getLatestReleaseZipUrl,
 	checkNestedZip,
 	downloadZip,
 };