+++ tests/e2e/specs/smoke-tests/update-woocommerce.js
@@ -3,7 +3,7 @@
  */
 const { merchant, utils } = require( '@woocommerce/e2e-utils' );
 
-const { getRemotePluginZip } = require( '@woocommerce/e2e-environment' );
+const { getRemotePluginZip, getLatestReleaseZipUrl } = require( '@woocommerce/e2e-environment' );
 
 /**
  * External dependencies
@@ -13,16 +13,23 @@ const {
 	beforeAll,
 } = require( '@jest/globals' );
 
-const { UPDATE_WC } = process.env;
+const { UPDATE_WC, TEST_RELEASE } = process.env;
 
-const nightlyZip = 'https://github.com/woocommerce/woocommerce/releases/download/nightly/woocommerce-trunk-nightly.zip';
+let zipUrl;
 const pluginName = 'WooCommerce';
 
 let pluginPath;
 
 utils.describeIf( UPDATE_WC )( 'WooCommerce plugin can be uploaded and activated', () => {
 	beforeAll( async () => {
-		pluginPath = await getRemotePluginZip( nightlyZip );
+
+		if ( TEST_RELEASE ) {
+			zipUrl = await getLatestReleaseZipUrl( 'woocommerce', 'woocommerce' );
+		} else {
+			zipUrl = 'https://github.com/woocommerce/woocommerce/releases/download/nightly/woocommerce-trunk-nightly.zip';
+		}
+
+		pluginPath = await getRemotePluginZip( zipUrl );
 		await merchant.login();
 	});
 
@@ -32,6 +39,11 @@ utils.describeIf( UPDATE_WC )( 'WooCommerce plugin can be uploaded and activated
 
 	it( 'can upload and activate the WooCommerce plugin', async () => {
 		await merchant.uploadAndActivatePlugin( pluginPath, pluginName );
+	});
+
+	it( 'can run the database update', async () => {
+		// Check for, and run, the database upgrade if needed
+		await merchant.runDatabaseUpdate();
 	});
 
 });