+++ tests/e2e/core-tests/CHANGELOG.md
@@ -14,6 +14,7 @@
 - The e2e test `order-status-filters.test.js` now uses the API to create orders
 - Implemented data-driven design using `describe.each` and `it.each`
 
+
 # 0.1.3
 
 ## Added