+++ tests/e2e/env/CHANGELOG.md
@@ -1,11 +1,18 @@
 # Unreleased
 
-- `updateReadyPageStatus` utility to update the status of the ready page
-- Added plugin upload functionality util that provides a method to pull a plugin zip from a remote location
+- `updateReadyPageStatus` utility to update the status of the ready page.
+- Added plugin upload functionality util that provides a method to pull a plugin zip from a remote location:
   - `getRemotePluginZip( fileUrl )` to get the remote zip. Returns the filepath of the zip location.
 - Added plugin zip utility functions:
   - `checkNestedZip( zipFilePath, savePath )` checks a plugin zip file for any nested zip files. If one is found, it is extracted. Returns the path where the zip file is located.
   - `downloadZip( fileUrl, downloadPath )` downloads a plugin zip file from a remote location to the provided path.
+- Added `getLatestReleaseZipUrl( owner, repository, getPrerelease, perPage )` util function to get the latest release zip from a GitHub repository.
+- Added `DEFAULT_TIMEOUT_OVERRIDE` that allows passing in a time in milliseconds to override the default Jest and Puppeteer timeouts.
+- Fix latest version tag search paging logic
+- Update fallback PHP version to 7.4.22
+- Update fallback MariaDB version to 10.6.4
+- Update fallback WordPress version to 5.8.0.
+- Remove unused WP unit test install script.
 
 # 0.2.2